refactor: addCardState에 cardName, setCardName 추가
