+++ src/pages/NameCardPage/NameCardPage.tsx
@@ -1,20 +1,32 @@
-import { useContext, useState } from "react";
+import { useContext } from "react";
 import { useNavigate } from "react-router-dom";
 import styled from "styled-components";
 import CardPreview from "../../components/CardPreview/CardPreview";
 import { GlobalContext } from "../../context/GlobalProvider";
 import { PAGE } from "../../constant/index";
 import { Button } from "../../components/common/Button";
+import { AddCardStateContext } from "../../context/AddCardStateProvider";
+import { Card } from "../../types";
 
 const NameCardPage = () => {
   const navigate = useNavigate();
 
-  const { cards, currentIndex, setCards } = useContext(GlobalContext);
-  const [cardName, setCardName] = useState<string>("");
+  const { cards, setCards } = useContext(GlobalContext);
+  const { cardCompany, cardNumber, expirationDate, ownerName, securityCode, password, cardName, setCardName } =
+    useContext(AddCardStateContext);
 
   const onSubmitHandler = () => {
-    cards[currentIndex].cardName = cardName;
-    setCards(cards);
+    const card: Card = {
+      cardName,
+      cardCompany,
+      cardNumber,
+      expirationDate,
+      ownerName,
+      securityCode,
+      password,
+    };
+
+    setCards([...cards, card]);
     navigate(PAGE.CARD_LIST);
   };
 
@@ -25,7 +37,7 @@ const NameCardPage = () => {
   return (
     <Container>
       <h2>카드등록이 완료되었습니다.</h2>
-      <CardPreview card={cards[currentIndex]} />
+      <CardPreview card={{ cardCompany, cardNumber, expirationDate, ownerName }} />
       <Form onSubmit={onSubmitHandler}>
         <Input placeholder={"카드이름을 지어주세요."} onChange={onChangeHandler} autoFocus />
         <Button isVisible={true}>확인</Button>
